--package-dir

+++ integration/src/test/scala/TestBase.scala
@@ -6,7 +6,5 @@ import scalaxb.compiler.xsd.{Driver}
 trait TestBase extends SpecificationWithJUnit with CompilerMatcher {
   val module = new Driver // with Verbose
   val tmp = new File("tmp")
-  if (tmp.exists)
-    deleteAll(tmp)
-  tmp.mkdir
+  if (tmp.exists) deleteAll(tmp)
 }
